requirements.txt
streamlit
pandas
gspread
google-auth
streamlit-authenticator
reportlab
pyyaml